pyproject.toml
[tool.poetry]
name = "eng-airflow"
version = "0.1.0"
description = ""
authors = ["Prog-LucasAlves <lucasalves_taz@hotmail.com>"]
readme = "README.md"
packages = [{include = "eng_airflow"}]

[tool.poetry.dependencies]
python = "^3.11"
apache-airflow = "^2.5.1"
apache-airflow-providers-postgres = "5.4.0"
pandas = "^1.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
